feat(telemetry): capture libcudf NVTX spans and serve NVTX routes (#1439)
## Summary

- Captures Sirius, libcudf, and CCCL NVTX events in both supported
deployments without a sidecar injection DSO:
- **Vanilla DuckDB + `LOAD`** exports Quent's initializer from the
Sirius extension and points NVTX at that DSO before the first NVTX call.
- **Statically linked Sirius** exports the same initializer from DuckDB
and maps a Sirius-private NVTX injection token to the running
executable, while forwarding every other `dlopen` unchanged.
- Honors `SIRIUS_DISABLE` and `enable_quent: false` by skipping
automatic injection; otherwise preserves configuration precedence:
`NVTX_INJECTION64_PATH`, then `sirius.telemetry.nvtx_injection_lib`,
then automatic discovery.
- Mounts Quent's NVTX catalog/viewport routes and pins Quent to
`381e1be0`, which includes
[quent#544](https://github.com/rapidsai/quent/pull/544) and the
`nvtx-sys` build fix from
[quent#697](https://github.com/rapidsai/quent/pull/697). Generalized
schema work remains tracked by
[quent#618](https://github.com/rapidsai/quent/issues/618).

## Verification

- `pixi run make`
- `pixi run make test` — 3,080 test cases and 32,833,137 assertions
passed
- `pixi run pre-commit run -a`
- TPC-H SF10 Q1 produced the exact same NVTX range/count signature in
both modes: 812 complete ranges across Sirius/default, libcudf, and
CCCL, with zero incomplete ranges or reconstruction anomalies.

---------

+++ rust/crates/telemetry/analyzer/src/tests.rs
@@ -60,12 +60,12 @@ fn tier_usage(resource_id: Uuid, bytes: u64) -> Option<Usage<MemoryTier>> {
     })
 }
 
-fn batch_event(id: Uuid, ts: u64, seq: u64, state: BatchPlacementTransition) -> Event<SiriusEvent> {
+fn batch_event(id: Uuid, ts: u64, seq: u16, state: BatchPlacementTransition) -> Event<SiriusEvent> {
     Event::new(id, ts, SiriusEvent::BatchPlacement(FsmEvent { seq, state }))
 }
 
 fn memory_tier_events(id: Uuid, parent: Uuid, name: &str, bytes: u64) -> Vec<Event<SiriusEvent>> {
-    let event = |ts: u64, seq: u64, state: MemoryTierTransition| {
+    let event = |ts: u64, seq: u16, state: MemoryTierTransition| {
         Event::new(id, ts, SiriusEvent::MemoryTier(FsmEvent { seq, state }))
     };
     vec![
@@ -145,7 +145,7 @@ fn fixture(with_batches: bool) -> Fixture {
     events.extend(memory_tier_events(disk_id, engine_id, "DISK", 16 << 30));
 
     // The query FSM: its first transition is the query epoch.
-    let query_event = |ts: u64, seq: u64, state: query::QueryTransition| {
+    let query_event = |ts: u64, seq: u16, state: query::QueryTransition| {
         Event::new(query_id, ts, SiriusEvent::Query(FsmEvent { seq, state }))
     };
     events.extend([
@@ -335,7 +335,7 @@ fn fixture(with_batches: bool) -> Fixture {
 /// - t=1000 exit
 fn add_working_space_task(fixture: &mut Fixture) {
     let task_id = fixture.task_1_id;
-    let task_event = |ts: u64, seq: u64, state: TaskTransition| {
+    let task_event = |ts: u64, seq: u16, state: TaskTransition| {
         Event::new(task_id, ts, SiriusEvent::Task(FsmEvent { seq, state }))
     };
     fixture.events.extend([
